Create the Link.js file and Link.css for each Link Tag. create the Links.js for all links to be shown.
Added the Inter font-family link to the index.html file. added the style to the index.css file body style. change the App.js File to a state function and included the state needed. modified the header.css file so that its margin = auto, this is to solve the edge display of image. add a div with classname imgSection to the Image.js. styled the ImgSection in the image.css file. pass props from the app.js state to the header.js from the header.js to the Image.js where i included it in the H2 of the Image.js file. styles the h2 file in the image.css file.